Create budget_app_project_freecodecamp.py
This is my fourth Python project. Half of the program I did fairly smooth and didn't see much of a difficult. But then when making the create_spend_chart function I couldn't get it right like the freecodecamp terminal wanted, turned out the problem was that I was using a class variable in which freecodecamp didn't allowed, I should calculate the total spending directly from the ledger. The whole process took me about 2 days with lots of debugging and tweaking in order to have the final result