Create polygon_area_calculator_project_freecodecamp.py
This is my 5th project. Tbh I think this project is the easiest one so far, it takes me only arounds 20 minutes! I got hardstuck in the get_picture method for a little bit because for some reason my terminal keep on printing the shape with the first row having additional space in the beginning. But after all, it was quite easy, I quickly ran through it with ease